requirements.txt
 # Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Async HTTP Client
httpx==0.27.2

# Environment Variables
python-dotenv==1.0.1

# Database (Phase 3)
# supabase==2.10.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0